tests/raft: deflake test_linarizable_barrier
The test waits until a barrier is attained until the dirty offset.

BOOST_REQUIRE_EQUAL(r.value(), leader_offsets.committed_offset);

A barrier doesn't guarantee that out of the box, calling it may return
an offset below dirty offset until which all hearbeats agree. If we
want a barrier until a desired offset, we have to ensure the barrier
returns >= desired offset, adjusts the calling code to do this.

Underlying reason is that flush_log() (called from the barrier)  maynot
always guarantee a flush, it may just return right away if there is
already a flush in progress (and hence nothing to flush). Meanwhile a
round of heartbeats guarantee barrier progress and the barrier returns
an earlier offset which doesn't include the dirty offset that pending a
flush.