cluster: track group id in shard_placement_table
We track group ids in partition assignments and current state because we
want to use them as kvstore keys - group id is just a number so it is
better in this regard than the ntp and it uniquely identifies the
current incarnation of the ntp (i.e. each group id is uniquely
mapped to ntp, but ntp can be mapped to several group ids, though only
one will be current).