c/configuration: cloud_storage_azure_hierarchical_namespace_enabled
This property is only meaninful for Azure, it's needed to decide if HNS
(Hierachical Namespace) is active for a storage account.

currently, if redpanda uses ABS, a REST endpoint is called at startup to
determine if HNS is active. This endpoint requires signed requests, so
cloud_storage_azure_shared_key has to be set and
cloud_storage_credentials_source has to be config_file.

To support cloud_storage_credential_source=azure_aks_oidc_federation or
azure_vm_instance_metadata, this property can be set to inform if the
HNS is active; this is needed because because these two credentials
source use OAuth instead of signing.

The property is an optional<bool>. if it's set,
abs_client::self_configure will return its value