storage: increase size of offset key map fragment size
With a fragmented_vector configured with 32KB fragments, the actual
fragment will be the largest (given T) that doesn't exceed 32KB. For the
offset key map the size of T=entry gave a fragment size of 20KB.

The default target size of the entire offset key map is roughly 128KB
which works out to a achieved capacity of 8192 fragments when adding
entries to the fragmented vector. Given that a vector is 24 bytes, this
means that the allocation was 8192 * 24 ~ 190K allocation and this
triggered the large allocation warning.

This PR changes the fragmented vector used in the offset key map to have
be a chunked_vector, which end up being around 80KB allocations when
fragment vector decides on the actual fragment size, so the final
allocation for the default configuration is around 45K.

Of course choosing a large map size would negate this solution. We don't
really expect large sizes to be configured. In any case, this should not
pose a problem is practice because we perform the entire allocation on
bootup when memory is not fragmented. If we observe large configurations
in the future we can continue to reduce the size (e.g. we could probably
bit pack some data in the entry to get a "better" fit of T=entry into
the target fragment size).

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit 425159322f60669a5bac1ba233943818cc989026)