redpanda/admin: add `Retry-After` to `redirect_exception`
This commit is dependent on commit 2cd9066533489742729f4fdc0ba047e0d601abc5
from https://github.com/WillemKauf/seastar, branch `http_exception_headers`.

In the case of multiple consecutive redirects in the admin server,
a `Retry-After` header is added to the http reply to the client.

This is done in order to avoid exceeding redirect limits when leadership
data is stale in the cluster and a request to the admin server is
made.

The URL for the `redirect_exception` now contains a `?redirect=` counter
in its query string, and every other redirect will have some back-off
appended to it for the next client-side request.