rptest: remove node from started list on kill like signal
This fixes a class of bugs like the one in
https://github.com/redpanda-data/redpanda/issues/17886.

We have a wrapper `RedpandaService.wait_until` which is used to wait for
conditions as long as the cluster is "healthy". Healthy is defined as
all "started nodes" being up.

In some tests, we do call `signal_redpanda` to kill a particular node on
purpose. In this case, we also need to remove the node from the started
nodes list otherwise the above check will trip thinking that a node is
crashed or something similar.

Fixes https://github.com/redpanda-data/redpanda/issues/17886