id_allocator: do not forward requests beyond first hop
A tight forward-to-leader loop has been discovered in a test where
metadata about leader is out of date:
https://github.com/redpanda-data/redpanda/issues/17873.

Instead, we remove the forwarding from the request handler and do it
only once on the original invoker. In
`id_allocator_frontend::allocate_id` we call
`allocate_router::process_or_dispatch` which will do the redirect and
retry if the target node returns an error/does not respond. It also has
backoff built in.

This is a fix very similar to one described in
https://github.com/redpanda-data/redpanda/pull/15953.

Fixes https://github.com/redpanda-data/redpanda/issues/17873