tests: Marking test_drain_audit* as ok_to_fail
- After investigation its been determined that the failure isn't really
an indication of auditing not working properly.

- The test verifies that on shutdown all stop application_lifecycle messages
are observed, and during shutdown there is a short grace period in which
shutdown will be held until these messages can be successfully produced,
however there is no guarantee that the produce will succeed as shutdown
is random and it may be the case that nodes that host the audit
partition to produce to have already been shutdown.

- Therefore the produce fails after the grace period and the message
fails to make it to the audit log, making the test fail.

- A potential fix is to have the audit log subsystem be partition
agnostic when producing messages, but this means it must have the
ability to  reacting to failures in a smarter way then just retrying,
currently our kafka client does not support this. Although attempts have
been made to implement this: https://github.com/redpanda-data/redpanda/pull/15202