rptest: fix KafkaCliTools.produce for SASL and TLS
KafkaCliTools.produce didn't work if SASL or TLS was enabled because
it failed to pass along the relevant configuration. This command does
not accept a config file via --command-config like many other cli tools
(this is how we pass this information to those tools), but it does
accept a --producer.config argument which serves the same purpose.

Set if additional security config needs to be passed in this way,
which fixes this test in the cloud.

Fixes redpanda-data/core-internal#1256