rptest: replace singal_redpanda with stop_node
We have a wrapper RedpandaService.wait_until which is used to wait for
conditions as long as the cluster is "healthy". Healthy is defined as
all "started nodes" being up.

When we call signal_redpanda to kill a particular node on purpose. In
this case, we also need to remove the node from the started nodes list
otherwise the above check will trip thinking that a node is crashed or
something similar.

I have tried to add this logic inside the signal_redpanda method in
[#17889][1] but discovered that there are tests which rely on the
existing behavior already.

@bharathv noticed that stop_node can be used instead so this PR does
that.

Fixes https://github.com/redpanda-data/redpanda/issues/17886

[1]: https://github.com/redpanda-data/redpanda/pull/17889