c/controller_backend: don't require cur_operation after reconciliation step
In case when we need to wait before shard_placement_table and
topic_table become consistent (i.e. the expected log_revision of an ntp
is the same), the concept of "current operation" doesn't make much sense
because we may not yet know what we are required to do for reconciliation.
Therefore we shouldn't require the current_operation to be non-null
after a reconciliation step returns an error.