rptest: fix test_exceed_broker_limit flake
ConnectionLimitsTest.test_exceed_broker_limit had a spurious failure in
CI. This test starts 2 consumers which (should) consume all 6
available connections, then checks that a producer started after that
fails to produce (due to connection limit being hit).

However the consumer & producer starts are all async, so the producer
can race ahead of one of the consumers and grab the connections for
itself, failing the test.

Change the test to wait for the consumers to connect, by waiting
until the connection metric hits 6, then starts the producer.

Fixes #17897.