c/pb_planner: randomly visit each replica (not each partition)
Previously, when doing counts rebalancing, we visited partitions
randomly, but for each partition we tried to move each replica in the
group sequentially. This can lead to undesireable clustering of
replicas - for example, if we add 3 new nodes, whole replica groups will
be moved there wholesale, and as a result, there will be no replication
traffic between old and new nodes. To avoid this, iterate over
individual replicas randomly.