tests/cluster_config_test: refactor in ClusterConfigAliasTest
ducktape creates a folder for a test run by appening the parameters to
the test name.

These test names are becoming longer than what's supported by the
OS, and can fail to start altogether.

This refactor passes in the name of the variable containing the
test parameter, and a new method fetches the actual variable from a
string