config/configuration: deprecate the name delete_retention_ms
before this commit, delete_retention_ms is accepted as an alias for
log_retention_ms.

this was done 1 release ago, and deprecation was reported in the
documentation.

This commits moves delete_retention_ms to a separate
deprecated_property.

it will be accepted but it will not cause side effects and it will be
not reported to the user when exporting the configuration