fetch: Don't format client ip:port for each partition
We were formatting the ip and port of the client for each partition in
the fetch plan. This is unneeded as it obviously doesn't change by topic
or partition.

The formatting is fairly expensive and amounts for ~2-3% of total
reactor time in a high partition count scenario.

This could be further optimized by moving the formatted version to the
connection context to avoid having this copy alive many times but that's
a bit larger change for another PR.