c/topics: use topic table revision to detect changes in generator
Topic table partition generator is a simple helper to asynchronously
iterate over a topic table which may have a lot of elements. The
generator validates if `topic_table` changed while the loop yield for
the other tasks. The check was incorrect as it was validating the topic
table revision which is updated after the snapshot is applied.

Fixed the issue by validating it `_topics_map_revision` changed during
the scheduling point.

Fixes: #core-internal/1259

Signed-off-by: Michał Maślanka <michal@redpanda.com>