rptest: upgrade key log messages to warning
Upgrade 3 key log messages in the cloud cluster creation workflow to
warning. These are not actually "warnings" but messages that we
definitely want to see in the part of the output that goes to stdout
(i.e., which are visible by default in the DT "session log"). For log
messages emitted by the tests themselves, this is only warning or
error messages so we use warning.

The session log/stdout is often all we get in the BK run when something
goes wrong.