rptest: Producer swarm updates
Add support for additional metrics, float values for p0, p50, p100
and waiting for all producers to start.

Also, allow swarm is_alive() to be 0-arg. Currently swarm always runs
on one node so it is weird to require the user to
pass "node" to this method, so remove that requirement.

Also, update producer_swarm.wait_node to bool.

In the case that the wait fails, i.e., we time out while waiting for
the node to stop we should return False, not propagate the
TimeoutException.