cloud_storage: Implement "carryover" behavior
The "carryover" behavior allows cache to use information from the
previous trim to quickly trim the cache without scanning the whole
directory. This allows cache to avoid blocking readers.

In a situation when the cache cntains very large number of files the
recursive directory walk could take few minutes. We're not allowing
number of objects stored in the cache to overshoot so all the readers
are blocked until the walk is finished.

This commit adds new "carryover" trim mechanism which is running before
the normal trim and uses information obtained during the previous fast
or full trim to delete some objects wihtout walking the directory tree.