dt/rbac_upgrade: Upgrade to head version instead of a specific 24.x
The intent of this test is to show that upgrading from a non-RBAC
supporting version to an RBAC supporting build creates the Users
role.

Previously, we would install v24.1.x explicitly. This broke because
we have tagged v24.2.0-dev (which bumps the head version) but have
NOT yet tagged v24.1.1 (so 24.1 doesn't appear in released_versions).

Simplest solution here is to upgrade directly to head version from
a non-RBAC supporting version, which should have the same effect
functionally but is more general.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>