c/controller_backend: bootstrap from shard_placement_table
When bootstrapping controller_backend, we don't really need to try and
reconcile all ntps in the cluster, we can start with the set that is in
shard_placement_table (i.e. all ntps with some relation to the current
shard). This also allows us to use non-consuming topic_table
notifications.