net: move unresolved_address to utils
net::unresolved_address poses a dependency challenge as long as it lives
in the net:: module. The reason it is that its use is prolific: the
nearly global modules of the configuration system uses and model module
uses it, and its used in numerous other subsystems.

Creating dependencies like config depending on net is not ideal.
Ultimately we may want to create specialized types just for the
configuration system. But until we solve those issues, likely when we
are complete or nearly complete with new style module transition, we
need a better place for unresolved_address to live, so we'll stick it in
utils.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>