tx/rm_stm: remove mem_state::estimated
'estimated' was used as an estimate for the first data batch offset (so
LSO cannot exceed it) while replication is in progress. This is the last
in memory state that is pending for cleanup.

This commit redefines the transacation boundary to begin from fence batch
rather the first data batch which allows us to get rid of this
additional state which lays the foundation for producer state move.