cloud_storage: throw if can not create manifest cursor
It is not reasonable to continue work after this point. The absence of a
cursor cursor is interpreted as EOF in other parts of the system. Not
throwing makes it impossible to differentiate between "no more data
available" vs. "an error occurred".

This is required for an upcoming commit that fixes a timequery bug where
cloud storage returns "no offset found" instead of propagating an
internal error.

(cherry picked from commit b53deac17c7f6bb2d3f1163965e4ef1397dafcac)