c/balancer_backend: first initialize planner and then call plan
This change is a part of an effort to identify and fix rare segmentation
fault in Redpanda that happens after it was suspended with `SIGSTOP`
signal.
According to the C++ standard the temporary should be kept alive until
the expression ends. The crash we are observing indicates the UAF issue.
The only way the variable, that access causes the segfault, can be
deleted is by getting out of scope which in this situation should be
guaranteed.

Given our experience with coroutines and different types of lifecycle
bugs that we found in past this is a poor man's effort to avoid the
issue.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit a52d0ada31dd0c842333fadc2efc37d4aaf2b471)