cloud_storage: do not ignore out of range errors
This code tried to be clever and ignore exception in some of the cases
where it was assumed it is safe to do so. I.e. if the start offset
stored in the cloud moved forward.

I don't believe covering these edge cases is necessary.
- Reasoning about correctness becomes harder as it masks the case where
  by mistake read an out of range offset.
- It hide from the client the fact that the offset they just tried to
  read does not exist anymore. As a user, if the log is prefix
  truncated, then I expect the reply to Fetch to be out of range and not
  an empty response.

(cherry picked from commit d1543eeb8c929eccb414da678445fa0981fca075)