admin: fixed mapping of service unavailable error code
Previously the service_unavailable error code was mapped to generic
unexpected server error. This lead to a situation in which a lot of
errors were logged when broker was restarted as it has an RPC server
running but without the services registered. Changed the mapping of
`rpc::errc::service_unavailable` to HTTP service unavailable error.

Fixes: #17739

Signed-off-by: Michał Maślanka <michal@redpanda.com>