cloud_storage: Run trim in the background
In case if carryover trim was able to release enough space start trim in
the background and return early. This unblocks the hydration that
reserved space and triggered the trim. We need to run normal trim anyway
to avoid corner case when the carryover list becomes empty and we have
to block readers for the duration of the full trim.

(cherry picked from commit 9f1b51b6a7e3455f3bc1c25965f3d19bb657dada)