rptest: kill existing connections in firewall_blocked
The intention of firewall_blocked is to always prevent communication.
However, if a connection already exists it might take a long time until
linux gives up retrying sending and the problem is reported. To avoid
this, we attempt to kill the connections instantly.

This was necessary for a variation of a timequery test I was writing.
Although is not strictly necessary anymore, I consider it to be a nice
addition.

(cherry picked from commit 4c706fb144441e1dfd1d030fda64cfd13cc8a9f5)