retention: conditionally enable retention for CO topic
This a special configuration some select users can use to enable
retention on CO topic because the compaction logic is ineffective
and they would like to use retention as a stop gap until that is fixed.
This configuration will be eprecated once we address the compaction gaps.
Should be untouched for all practical purposes.