rm_stm: fix fence_pid_epoch cleanup
fence_pid_epoch maps a producer id to its latest epoch. Current cleanup
code does not do a epoch check before cleaningup the pid state. This can
result in removing the state related to the latest epoch. Consider the
following series of events..

[x, y] = pid[id=x, epoch=y]

[1, 0] begin_tx - fence_pid_epoch[1] = 0
[1, 1] begin_tx - fence_pid_epoch[1] = 1
evict [1, 0]
erase(fence_pid[1]) ==> removes (1)

This results in a messed up state stalling the state of the transaction
because the partition cannot make progress until it verifies the epoch.

This is a long pending bug that was exposed by racy evictions.

(cherry picked from commit 996e138)