storage: add min/max bounds for timequeries
kafka ListOffsets request allows to query offsets by timestamps. The
result of such a request is the first kafka offset in the log that has a
timestamp equal to or greater than the requested timestamp. Or, -1 if no
such record can be found.

The implementation we have today assumes that the start of the physical
log matches the start of the log as it is seen by external users/kafka
clients.

However, this is not always true. In particular, when [trim-prefix][1]
(prefix truncation) is used. There are 2 sources of problems:

  1) trim-prefix is synchronously applied at cluster layer where it
     changes the visibility of the log from the client point-of-view,
     but it is asynchronously applied to the consensus log/physical log/
     disk_log_impl class, cloud storage.

  2) trim-prefix is executed with an offset of a record that in in the
     middle of a batch.

As a result, in these scenarios, if the clients sends a kafka Fetch
request with the received offset they'll be replied with
OffsetOutOfRange error.

This commit changes such queries are implemented at the lower levels
of the system by carrying the information about the visible start and
end of the log together with the timestamp. Then, at the lower levels,
we use these to limit our search only to that range.

Although this commit does not change the implementation of the tiered
storage timequery it does fix the trim-prefix problem there too in the
general case because of check and "bump" added in
https://github.com/redpanda-data/redpanda/pull/11579.

Tiered Storage timequeries have some additional problems which I plan to
address in https://github.com/redpanda-data/redpanda/pull/18097.

[1]: https://docs.redpanda.com/current/reference/rpk/rpk-topic/rpk-topic-trim-prefix/

(cherry picked from commit 76a1ea2452b09a5730f2574646fc06ab2b8b8e32)