tests: adjust time between failures in relaxed consistency test
Recently introduced Raft change that allows a candidate to wait for
replies from all the nodes in order to detect the one with longest log
may make the leader election a bit longer. Change the interval between
failure injections in the test to allow nodes to recover and elect the
leader.

Signed-off-by: Michał Maślanka <michal@redpanda.com>