cst: download chunks from bg loop
Remote segment background loop now downloads all enqueued chunks. This
download was previous done by chunk API from any fiber. The downloads
are only performed at the end of a loop iteration, provided that the
index is materialized and we are not operating in legacy mode, where
full segments are downloaded.

The chunk download requests are kept in a fragmented vector. Only one
download is expected per chunk start offset. The first request for a
chunk is added to the vector, the chunk API is expected to enqueue the
subsequent callers into a waiter list. Since chunk consumers use a
shared ptr to access the chunk file handle, and the file handle is
evicted when there are no consumers attached to it, we maintain a
single entry for a chunk download request within remote to ensure that
we do not return multiple ss::file objects for the same chunk file.