cloud_storage: Improve logging and error handling
Improve logging and error handling in the remote_partition.
In the remote partition reader we're checking two abort_source
instances. One belongs to the remote_partition instance and another
belongs to the sharded_abort_source that belongs to the kafka request
ctx. The last abort source is used to propagate exceptions from kafka
layer. The abort_source in the remote_partition is not used for this
purpose. It's only used to shutdown and therefore its 'check' method
only throws 'abort_requested_exception'.

Because of that any kafka disconnection error triggers error level
logging in TS. This commit changes behavior a bit. In case if the
exception from the kafka layer is triggered by the disconnect the error
is converted to simple shutdown error.

(cherry picked from commit 8502a4d78dd8e410309fc6d258573be63b4250f8)