mvlog: skipping data source
Adds a data source that wraps the paging data source that allows a
caller to pass in a set of intervals to read through. This will be
useful when reading a segment that has a gap in it caused by truncation.

While the only real dependencies here are on the io:: subsystem, I put
it in storage::mvlog since that's where it will primarily be used.

This also uses the pager directly, which is subject to change as the
io:: public interface evolves.