archival: live settable manifest upload timeout
This is needed for a ducktape test where we want to change the manifest
upload timeout at runtime. E.g. set it to 0 to prevent manifest
uploading from on point in time onward.

It is declared in configuration.cc as not requiring restart but in fact
it did require one prior to this commit.

Other properties of the archiver configuration should be fixed too in a
separate commit.

(cherry picked from commit aab5fe7db7c96d1623a1e56e5263f887a5016fb9)