archival: live settable manifest upload timeout
This is needed for a ducktape test where we want to change the manifest
upload timeout at runtime. E.g. set it to 0 to prevent manifest
uploading from on point in time onward.

It is declared in configuration.cc as not requiring restart but in fact
it did require one prior to this commit.

Other properties of the archiver configuration should be fixed too in a
separate commit.