c/shard_placement_table/ut: assign shards concurrently
For now, when assigning shards only involves modifying in-memory state,
we can get away with assigning each ntp sequentially. But in the future
it will also involve kvstore writes (with their associated 10ms
latency), and for the case of 1000s of partitions this will lead to
noticeable delays. Assign shards concurrently to deal with this problem.