model: add bounded_offset_interval type
Encapsulates common operations on offset intervals. For now, although it
is named bounded, the maximum offset can still be set to
`model::offset::max()`. I will likely change this in the future as it
requires changing quite a bit of call sites, most likely only tests.

This little data structure tries to be very light weight and impose
minimum overhead on basic interval operations like intersection or
inclusion. It is also quite hard to use it incorrectly due to the
checked construction variant and debug assertions.

Later, we might want to refactor things like log_reader to use this
instead of min and max offsets like they do today. Once that is done,
the checked variant needs to be called only once at the kafka layer. For
everyone else it becomes a ~0 cost abstraction.