cloud_storage: carry min/max offset bounds into cloud storage timequery
Tiered Storage physically has a superset of the addressable data. This
can be caused at least by the following: a) trim-prefix, b) retention
being applied but garbage collection not finishing yet.

For offset queries this isn't problematic because the bounds can be
applied at higher level. In particular, partition object does validate
that offset is in range before passing control to the remote partition.

For timequeries prior to this commit such bounds were not enforced
leading to a bug where cloud storage would return an offset -1 (no data
found) in result when there actually was data or returning a wrong
offset.

Wrong offset: it would be returned because reads could have started prior
to the partition visible/addressable offset. E.g. after retention was
applied but before GC was run. Or, after a trim-prefix with an offset
which falls in a middle of a batch.

Missing offset: would be returned when the higher level reader was
created with visible/addressable partition offset bounds, say \[1000,
1200\] but cloud storage would find the offset in a manifest with bounds
\[300, 400\] leading to an out of range error which used to be ignored.

(cherry picked from commit 0735bdfdabf8dde7b95303e9b9b28d8851d9eb9e)