cluster/self_test: use prefix in `verify_list()`
In the case that `cloud_storage_enable_remote_read` was true and a
payload was uploaded, the `verify_list()` test in the cloud storage
cluster self-test will attempt to list from the `self_test_prefix`
and later assert that the uploaded payload was found in the results.

If the payload was not uploaded, the prefix will be `std::nullopt`
and the root of the bucket will instead be listed from, as before.