tx/rm_stm: deal with duplicate begin_tx requests
Chaos logs showed that there can be duplicate (back to back) begin_tx
requests due to some certain client quirks beyond our control. We had
some checks in place to ignore them but they were invalidated by
https://github.com/redpanda-data/redpanda/pull/18076 which redefined the
transaction boundries.

After #18076 an entry in the ongoing_map does not mean the data has been
replicated as a part of the transaction. This PR adjusts the checks
accordingly.

Note: This code is going away soon, so we will have better API in the
producer_state to make the code more easily understandable. This PR is
to make chaos happy until then.