cloud_storage: add `continuation_token` parameter
This commit adds an optional `continuation_token` parameter to
`remote::list_objects()` in order to support future work around
re-factoring of the `list_objects()` remote API.

This allows users to handle results that are truncated at the call site
to `list_objects()`. The `remote::list_result` will have `is_truncated`
set to true, and `next_continuation_token` set appropriately. To allow use
of this `next_continuation_token` in future requests, we have to expose
it as a parameter in `list_objects()`.