mvlog: wrap pager with file abstraction
Adds a barebones abstraction around the io:: paging abstractions that
looks more like a file, with the goal in mind to abstract away the
details of the underlying io:: subsystem.

For now, this file abstraction supports appends, reads, and various file
lifecycle operations, which is enough to get started using it in
development of the new local log.