storage: rename maybe_override_retention_config()
The previous name is pretty vague: as a reader it's unclear what it's
actually doing other than modifying the config in some way.

Updates it to explicitly mention that it is handling local storage
overrides. IMO this is easier to reason about when reading the method
and its call sites.