cst: Handle remaining waiters on segment stop
Chunk waiters are stored in a list which is swapped before processing.
Because chunk downloads take time, it is possible that new waiters may
be registered while some downloads are being processed. The remote
segment may receive a stop request during this interval. If this
happens, the hydration loop will shut down, not processing the
remaining waiters, and the waiters will prevent gate closure of remote
segment, causing a hang.

The workaround cancels pending downloads while exiting the background
loop.