cst: Implement background loop in chunk API
The chunk API is now structured around a background loop similar to the
remote segment. When a download request for a chunk arrives, it is
added to the wait list for the chunk. Then a condvar is triggered,
which wakes up the bg loop and request is issued to remote segment to
download the chunk.

Once the chunk is downloaded, all waiters are notified, which are also
notified when there is an error during download.