resource_mgmt: add clarity to info+ logs
We have seen issues where the lack of tiered storage resulted in disk
filling up. The info logs didn't make it clear what was happening and
it was assumed to be a bug in code, rather than a misconfiguration
issue. This commit bumps a log line to info level so that it would have
been clear that a specific error path wasn't being taken.

Also tweaks wording to makes a stronger connection to tiered storage.