cmake: fix prevent seastar changing to c++23
Seastar uses a cmake cache variable to configure the c++ version it
uses. The problem with this for the redpanda build is that the cache
variable overwrites the normal variable set with the simple set() call.
This leads to us building redpanda with c++23 instead of c++20 as
intended, since seastar started setting the c++ version to c++23.

This is only a problem with the open source build because here we use
FetchContent to include dependencies directly in the cmake build,
whereas the internal build uses ExternalProject which allows us to
explicitly set the cache variable `CMAKE_CXX_STANDARD` for our
dependencies.

Note that we can't yet change to compiling with C++23 because of a bug
in the llvm-project. The combination of libstdc++ >= 13, c++ >= 23 and
clang 16 has a bug that caused compilation issues on ubuntu (see
llvm-project issue 61415).

This does not cause any compilation errors on fedora because fedora has
a patch version higher gcc and libstdc++ installed which contains a fix
for the above bug.