archival/tests: yield more often to the scheduler
Adding a segment takes ~1ms. Adding multiple segments in a single batch
can stall the reactor for a while. This, together with a recent seastar
change[^1] caused some timeouts to fire very frequently[^2].

Fix the problematic test by splitting the batch passed to add_segments
so that we execute more finer grained tasks and yield more often to the
scheduler.

[^1]: https://github.com/scylladb/seastar/issues/2238
[^2]: https://github.com/redpanda-data/redpanda/issues/13275