http: do not retry connection errors
I spotted that in some rpfixture tests the cluster metadata uploader is
running with an incorrect port[^1]. Beside logging lots of connection
refused errors, I found it to adding a non-negligible overhead to some
variations of the test. The overhead is result of relatively tight loop
in the `client::get_connected` method which I'm trying to remove in this
commit.

I don't believe this is necessary.  Returning the error to the caller is
much better. The callers usually have better retry mechanisms in place
with backoff and can log much more useful messages since they have
access to more context about the operation.

[^1]: We need to investigate separately why it doesn't get the updated
configuration when we configure cloud storage (you can try this test for
example https://github.com/redpanda-data/redpanda/issues/13275) or if it
needs to run as well.