storage/tests: Adjust step size deviance used in test
In the e2e storage test the index step size of 32KiB is used as possible
deviance from expected results. An allowance of 50KiB is used to prevent
the test failing because of index step size.

The change computes max step size across all segment indices and uses
that as the deviance.