tests: wait for messages before adding a node to cluster
In the `test_decommissioning_rebalancing_node` test case we check if a
node that joined the cluster and have some partitions assigned to it in
the process of data rebalancing can be successfully decommissioned.

The test was flaky as sometimes all the partition rebalance actions were
finished before we validated if rebalance started.

Added a condition to make sure that we wait for more data before adding
the node to cluster to make sure rebalance will last long enough for the
decommission to interrupt it.

Fixes: #13522

Signed-off-by: Michał Maślanka <michal@redpanda.com>